adding showpopup.js module
